keep deployment imports free of httpx when serializing chat tool arguments
